ai_liaotian/requirements.txt
Flask==2.3.2
Flask-Login==0.6.2
python-dotenv==1.0.0
requests==2.31.0